Add support for systemd 216 and up.
systemd 216 replaces the 'Servers=' configuration option in
timesyncd.conf with 'NTP=' and added 'FallbackNTP='.
This commits add version detection of systemd and updates the template
to adjust the configuration accordingly.

Also updated the meta information with more distributions this should
work for and increase the minimal Ansible version to 2.5. I had to raise
the minimal Ansible version because of the Jinja2 `version` test within
the template.