Add FallbackNTP functionality added by systemd 216.
This commit updates the template to add the FallbackNTP= option.
Configuration variables have been altered accordingly. `timesyncd` is no
longer a dictionary with an embedded list. Each variable is now it's own
distict variable to allow for flexibility and overriding the defaults in
a simple manner.

Also overhaul README.md and update meta.